video_pipeline/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI Models
openai==1.3.5
openai-whisper==20231117

# HTTP
httpx==0.25.1
aiohttp==3.9.0

# Video/Audio Processing
opencv-python==4.8.1.78
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0

# Image
pillow==10.1.0

# Utils
python-dotenv==1.0.0